Change mutation id from int to float and changed the findOrFail parameter to simply id.

+++ admin/src/graphql/confirmPendingCreation.js
@@ -1,7 +1,7 @@
 import gql from 'graphql-tag'
 
 export const confirmPendingCreation = gql`
-  mutation ($id: Int!) {
+  mutation ($id: Float!) {
     confirmPendingCreation(id: $id)
   }
 `
